+++ changelog.md
@@ -2,3 +2,4 @@ Codion Change Log
 ==================
 
 ## 0.17.18-SNAPSHOT
+- Text.padString() deprecated for removal along with Text.ALIGNMENT, Text.leftPad() and rightPad() added.